requirements.txt
flask
openai
langchain
faiss-cpu
mysql-connector-python
python-dotenv
